Add docker-compose for nginx, php, pgsql
This is the initial attempt at setting up a Docker environment for the
project. I've never done this before, so there might need to be some
significant updates to these environments as we go (for example, I
imagine there will be some configuration options passed to nginx to
allow proper routing to our front controller, as well as support for
TLS). But this is a foot in the door!